specify two integral potentials
