+++ src/potentials/potentials.jl
@@ -7,19 +7,32 @@ Abstract type for the considered potential
 abstract type Potential end
 
 """
+Abstract type for potentials that depend only on r (spherically symmetric or perfectly flat)
+"""
+abstract type TwoIntegralPotential <: Potential end
+
+"""
+Abstract type for potentials that depend on r,z (flattened)
+"""
+abstract type ThreeIntegralPotential <: Potential end
+
+"""
 Abstract type for central potential
 """
-abstract type CentralPotential <: Potential end
+abstract type TwoIntegralCentralPotential <: TwoIntegralPotential end
+abstract type ThreeIntegralCentralPotential <: ThreeIntegralPotential end
 
 """
 Abstract type for central potential with finite value in the center
 """
-abstract type CentralCorePotential <: CentralPotential end
+abstract type TwoIntegralCentralCorePotential <: TwoIntegralCentralPotential end
+abstract type ThreeIntegralCentralCorePotential <: ThreeIntegralCentralPotential end
 
 """
 Abstract type for central potential with infinite value in the center
 """
-abstract type CentralCuspPotential <: CentralPotential end
+abstract type TwoIntegralCentralCuspPotential <: TwoIntegralCentralPotential end
+abstract type ThreeIntegralCentralCuspPotential <: ThreeIntegralCentralPotential end
 
 #####################################
 # Generic functions (not methods)
@@ -29,7 +42,7 @@ abstract type CentralCuspPotential <: CentralPotential end
 
 Potential value for central potential `model` at radius `r`.
 """
-function ψ(r::Number, model::Potential)
+function ψ(r::Number, model::TwoIntegralPotential)
     # ... [model specific implementation] ...
 end
 
@@ -38,7 +51,7 @@ end
 
 Potential derivative value for central potential `model` at radius `r`.
 """
-function dψ(r::Number, model::Potential)
+function dψ(r::Number, model::TwoIntegralPotential)
     # ... [model specific implementation] ...
 end
 
@@ -47,7 +60,7 @@ end
 
 Potential second derivative for central potential `model` at radius `r`.
 """
-function d2ψ(r::Number, model::Potential)
+function d2ψ(r::Number, model::TwoIntegralPotential)
     # ... [model specific implementation] ...
 end
 
@@ -56,7 +69,7 @@ end
 
 Frequency scale associated to the potential `model`.
 """
-function frequency_scale(model::Potential)
+function frequency_scale(model::TwoIntegralPotential)
     # ... [model specific implementation] ...
 end
 
@@ -65,7 +78,7 @@ end
 
 Energy scale associated to the potential `model`.
 """
-function energy_scale(model::Potential)
+function energy_scale(model::TwoIntegralPotential)
     # ... [model specific implementation] ...
 end
 
@@ -74,7 +87,7 @@ end
 
 Angular momentum scale associated to the potential `model`.
 """
-function momentum_scale(model::Potential)
+function momentum_scale(model::TwoIntegralPotential)
     # ... [model specific implementation] ...
 end
 
@@ -83,15 +96,19 @@ end
 
 Radial scale associated to the potential `model`.
 """
-function radial_scale(model::Potential)
+function radial_scale(model::TwoIntegralPotential)
     # ... [model specific implementation] ...
 end
 
 #####################################
 # Include potentials
 #####################################
+
+# two integral potentials
 include("isochrone.jl")
 include("plummer.jl")
 include("hernquist.jl")
 include("mestelzang.jl")
+
+# three integral potentials
 include("kuzminkutuzov.jl")
